feat: add more time jmespath filters
